Add suportability for Shared mailbox
Some customer asked us about OME/encryption with SharedMailbox. There's no way to open such items.

From existing bug, there's comment from dev about this is known issue and it's listed to feature backlog.
https://o365exchange.visualstudio.com/IP%20Engineering/_workitems/edit/812810
They issue with OME in OWA for shared mailboxes a known issue for them and is already added in the backlog for fixing, testing and implementation.
Currently OME is not supported for Shared Mailboxes.
Unfortunately I cannot give you an ETA, when this feature will be implemented.